// library-backend/src/modules/admin/audit-logs/audit-logs/interfaces/audit-logs.interfaces.ts
export interface IAuditLog {}

export interface AuditLogsLogActionParams {
  entity: string;
  entityId: string;
  action: string;
  oldValues?: Record<string, any> | null;
  newValues?: Record<string, any> | null;
  performedById?: string;
  performedByName?: string;
  performedByRole?: string;
  tenantId?: string;
  branchId?: string;
  ipAddress?: string;
  userAgent?: string;
}